feature: separated bootable logic into trait
